Altered a few labels to make their meaning more clear.
The button to submit a rating is now named Rate instead of Vote, the button to exit the list of reviews is now called Close instead of Back. Also added an introduction text to the reviews list. It suggests to submit a review if there is none. Also added the default option to sort the add-ons by rating to the add-on filter options.