Upgrade Xcode project to enable both i386 and x86_64 builds
The largest part of this consist of upgrading the boost and pango/cairo
libraries together with their dependencies, supporting files and
project settings. In addition, architecture, SDK and compiler options
are set to use the native versions of the given OS X version for debug
builds. The OS X deployment target is set to OS X 10.8, because of a
change in the default libraries used by OS X from 10.8 to 10.9.
For this to work, the resources in  wesnoth_compile_mac_1.13.zip need
to be installed as well.
Some other settings, e.g. font defaults, still need to be adjusted.
This will be done in separate commits. Settings for release builds will
be changed later as well.