enable undo recruits/recalls if no rng was used
If we used rng during an acion we can never undo it becasue it is an nonundoable information gain. Also it's technicaly impossible because in mp using rng will result in data send over the network, which cannot be undone. Wedon't need to check whether the unit had traits/names this was becasue that would also have caused rng calls.