provide a desktop notification when game is ready to start
this is played at the same time as the turn bell, it is effectively
the same as a chat message which will cause a desktop popup on
systems that we support this for.

note: we don't check notification preference, because this should
be done in the show_notification function, not the caller.
however, it seems that is only done with DBUS notifications? why
not others?