add [start] tag to server replays as needed by gfgtdf
this should have been part of 121, but wasn't
digest from irc logs:
post PR 121, all [replay]'s are expected to begin with an item
[command]
    [start]
    [/start]
[/command]
which may also contain some [checkup] children which help to
ensure that we didnt go OOS in prestart.
these commands are not generated by the server and distributed,
because we want the clients to begin loading the game without
waiting for the server to finish.
all clients, including observers, will add them automatically.
the server therefore should not add them to the history, and
rather just prepend them to [replay] when generating the
savefile.