fix bug #21916, ready blocked at inappropriate times
the problem was that params.saved_game was checked at the wrong time,
it was checked at the time of making a new config, but not in the
constructor. the result is that the host and clients are out of sync
about whether allow_changes is true for a given side.