always allow SAVE_REPLAY
We diallow save game during action because that would lead to broken saves, but this is not true for save_replay.
If, for some reason, we get stuck inside a an action we cannot save the game, with this we can at least save the replay in this case and create a vaild save with the 'replaying + saving durign replay' strategy.