LoW: Initialize units in start events
previously this was done in side 1 turn 1 events because start events
wasn't synced.

After the fix for 21933 players now loose when they have no units after
start events, so we need to do this in a start event.

replacing check_victory with check_end_level after start events in
play_controller.cpp line 572 doesn't work because we still have the
check_victory in playmp_controller when waiting
for remote packages, that check_victory is needed in order to sync the
behavior of ai turns which calls check_victory after every action on the
local client.