fixup issue in 21939
the issue is that linger mode makes things very complicated for the
server to handle correctly. when some players are lingering on the
old level, and some begin to play on the new level, we have to keep
the command messages and various server control messages seperate
or we get errors.

it is more of an issue than it was before because with the server
controller tweaks, the server will immediately send controller
change messages when the level transition occurs.

the most obvious "bandaid" without removing code or changing the
architecutre is to track which players have advanced or not (and
whether this was a game where we advanced) and filter out messages
as appropriate. that is the effort that i began here.

alternative fixes include:
- use a new game object for the level we transition to.
- disable linger mode in networked mp
- make it so that lingering players transition as soon as the host
does.

discussed on irc briefly april 20 2014