partly revert e454252f218b
this should restore the behaviour that players don't loose if they have
no units left after the start event. Because it wasn't tested enough.

Note, that in most cases we don't need the check_victory in
playmp_controller::play_network_turn because we have a check_victory()
in the end in run_in_sycned_command