fix 21903
it's now impossible for non hosts to advance to the next scenario before
the host does, this is the intended behavior which didn't work before
because the hotkeys system automaticly sets the endturnbutton enabled if
can_execute_command(turn_info::HOTKEY_ENDTURN) == true.

We also make the clients end linger mode immediately when they receive a
PROCESS_END_LINGER, to prevent receiving data of the next scenario while we are not playing it.