more robust replay.cpp functions
we now add nonundoable commands at the current position instead of
assuming at_end(), this way nonundoable replay entrys like speak
can now also be added while we are not at end.
Usually playturn_network_adapter ensures that we are always
are at_end() and i didn't get any error with the previous solution
but the new solution is just more robust in case something goes wrong.

We cannot just do this for normal commands because it might
bring the replay in disorder.
We cannot add nonundoable commands at end if we aren't at end
because in that case the user would see the message again in
do_replay_handle in a mp game.