add mt_rng to server for seeding mp games
The server holds an mt_rng for every game, which it uses to
generate random seeds which are sent to the clients. This way
there is only one code path to boost::mt19937 and
boost::random_device.

TODO: It looks like this will still cast all uint32's from the rng
to ints before sending them to the clients, unclear if that causes
problems.