Test sending rng seeds over the network
This test makes a series of configs which look like rng seed updates,
sends them using the network layer, and recieves them, and checks the
integrity. It includes hex strings which are all numeric and begin
with leading zeros, to be sure that the leading zeros are not
truncated. The list of seeds is duplicated from part of test_rng.cpp.