redirect noninteractive ai test info from std::cout to log channel
I named the log channel "aitesting" and in three files,

playcontroller.cpp,
play_singlecontroller.cpp,
ca_testing_recruitment.cpp,

bound this to "LOG_AIT" and substituted where I saw std::cout.

This macro could be redefined to std::cout to restore previous
behavior.

Feel free to rename the channel if you don't like this name.
The reason I made this change is so that this info does not
appear when we are running noninteractive unit tests.