add wml commands 'issue_attack_command', 'issue_move_command'
This is the initial commit for these WML handlers.

The purpose of these is to provide more or less raw interfaces
to the (synchronized) implementations of attacks, moves, used
when a user clicks to make such an order, or when an ai issues
such an order.

This is useful for creating noninteractive tests, as it allows
to test the outcomes of these kinds of things without having
to figure out how to coerce an ai to do them. Sure, you can
make the same effects other ways in WML, but the point is
to test the code path that is actually used in a game.