add unix timeout option to wml test script, enable empty_test
Adds an option to the new wml test script, to use the unix
timeout command instead of wesnoth's built-in one. This is because
the built-in one is very slightly unreliable... there's a reason
killthread was deprecated in SDL unfortunately. Will leave it for
UMC but on travis will replace with unix timeout.

The unix timeout one seems consistent enough that I will enable
the empty_test timeout check for travis.

Also, adds a help listing to the wml test script.