add [do_command] wml action tag
[do_command] executes commands like attack, move, recruit, etc.,
similar to those that may be found in [command] tags in replays.
These are executed in game immediately in a synchronized context,
similarly to as if a human or ai controller had ordered them.

It is also possible to use the command to fire a WML game event,
or to execute lua code from a synchronized context.

Only some of the children of [command] are valid in this tag.

With this tag, [issue_attack_command] and [issue_move_command] are
both deprecated and will be removed shortly.

Thanks to gfgtdf for both the idea and the implementation.