in debug mode, add team strings as well as names for hidden sides.
This brings the team names in line with the player names and other
behavior in mp_connect in debug mode. For instance there are other
lines like this:
https://github.com/wesnoth/wesnoth/blob/dbf3331e9a99e75ca0c16009169eebeddb9c1db1/src/multiplayer_connect_engine.cpp#L155

From inspecting the save files generated while testing this issue,
it's not clear if it could have actually caused a problem in game,
but it addresses an issue raised by SlowThinker on the bug tracker:

https://gna.org/bugs/index.php?21800

and does indeed fix a minor shortcoming of debug mode in mp connect
engine.