Move animationless moving unit hiding to a more appropriate place
Unit hiding is usually done in replace_temporary, which is called by
proceed_to. As proceed_to exits at the beginning if animate is false, this
would seem like a good place to hide the unit, but it is also unhidden during
this function. This means that was_hidden_ is set multiple times, with the
second time overwriting it with the temporary state of being hidden.
As animate remains false for the entire lifetime of the unit mover, we hide it
at the start and let finish unhide it again.

Reported at http://r.wesnoth.org/p570866