remove unnecessary variable "turn_over_" from play_controllers.
After other refactors, this variable is basically useless -- it
is only used in one function, and indicates to that function if
we called it from the normal sp path, or the mp linger code path.
We choose to replace it with the linger_ variable -- the sp code
path will never follow to this function if linger_ is true, and
if it did it would cause a serious error anyways and corrupt
replays, so we would know if this was happening. So this is a
strict refactor.