use completion setter object in mp linger mode
mp linger mode was not properly setting the completion status,
as sp does. This caused mp end of scenario saves not to end up
in linger mode when they are reloaded.

Note that this reopens bug 15545 (!)
https://gna.org/bugs/index.php?15545

In 1.11+, I believe that the correct way to fix that is to save
the [carryover] while in linger mode, and restore it on reload,
however this is for another commit...