replay controller uses game_board to do resets
This is not a strict refactor -- the game map will also be reset
to the starting configuration now, when we reset the replay. It
doesn't look like this was the case in the old version, so it's
possible that if a scenario used [replace_map] or similar, the
replay could go out of sync after hitting the reset button.

Regardless of whether the old version was buggy, the new version
should be just as good if not better, and was tested.