remove resources::state_of_game
we don't want to expose the fileds "snapshot", "carryover_sides".
instead we only add classification, and mp_game_setting to resources
which are members of game_state.

in order to do that, we have to add a do_autosave method to
play_controller.

this commit also moves some code from game_state::write_snapshot to
play_controller::to_config which was intended to be in a different
commit.