main catches and handles "std::exception" and "...", returning 1
The purpose of this is so that even when we request the program
to be terminated with "throw 42", the exception will still
propogate back to main and all destructors will be called along
the way. This way, if the program requested a socket for example,
the socket will still be released when we terminate unexpectedly,
so it might be slightly more convenient for testing.