Fix an assertion failure in the GUI2 placement.
If a vertical scrollbar is added to a widget its size is recalculated.
Unfortunalty the determination of the minimum width didn't take the
horizontal scrollbar in account. This lead to the container to be too
small for a widget with horizontal and vertical scrollbar causing the
placement to fail.

Fixes bug #22095, possibly #15615 and #21785.