Fix an assertion failure in the GUI2 placement.
If a vertical scrollbar is added to a widget its size is recalculated.
Unfortunalty the determination of the minimum width didn't take the
horizontal scrollbar in account. This lead to the container to be too
small for a widget with horizontal and vertical scrollbar causing the
placement to fail.

Fixes bug #22095, possibly #15615 and #21785.

There are still a few issues remaining, but it seems they only cause the
code to behave less efficient. The fix for this issue seems to be more
intrusive and will only be committed to master.