gui2/tchat_log: Add a button to copy the filtered page contents to clipboard
This copies the current contents of the dialog to clipboard. The button
currently lacks a tooltip because the tooltip has the potential to cause
map labels to glitch through the dialog when displayed (see commit
eab3e6fb646fda8cc6101a3e568c86c2b17b707f and bug #22176).

(The tooltip should also remain commented-out because this is a backport
to 1.12.)