new_release: Workaround race condition in wesnothd SIGHUP handler
If we try to send a message immediately after signaling SIGHUP to
wesnothd, it won't get caught by wesnothd, probably because the control
socket is being reopened just as the message arrives. Thus, we need a 1
second delay after the signal.

This was fixed in production a while ago but I forgot to commit the
change.