Don't call preprocess_file() on an empty path if fonts.cfg can't be found
get_wml_location() returns an empty string on failure, and there is
really no point in calling preprocess_file() on that -- it only results
in a confusing error message.

Replaces this sequence often found when Wesnoth can't resolve the path
to its data directory:

  20140612 23:37:05 error filesystem: Trying to open file with empty name.
  20140612 23:37:05 error config: Could not open file
  could not initialize fonts

... with the following one:

  20140612 23:38:52 error font: could not resolve path to fonts.cfg, file not found
  could not initialize fonts

... which should be a little bit more helpful in pointing users in the
right direction in the aforementioned case.