fix random map generation
this was previously broken because we passed a temporary pointer to
generate_map. While fixing this i moved it to saved_game.cpp

i also added a call to expand_random_scenario in mp_game_utils so that
in a mp campaign players advance to the next scenario which is randomly
generated. they have all the same scenario.